requirements.txt
PyYAML
pywebostv
pynput
